pkd: break out early in final `!ssh_is_connected` loop
When testing with dropbear I observe that it is likely that the
loop removed here will encounter an error because the channel
at hand has already been closed by the dropbear client-side.

To fix the problem, break out early if `ssh_is_connected` is
no longer true upon completion of the poll.

The loop removed here intends to gracefully wait for the
connection to close; it can be removed without affecting the
validity of the test.  Connection state will be cleaned up in
the `goto out` exit path.

Signed-off-by: Jon Simons <jon@jonsimons.org>